Added new icons. Implemented lever icon control on main page and cover. Wrote SettingsPage. Wrote settings C++ object QML interface. Successfully reads & sets settings from the app, but something is calling startTransmission even though it isn't called from the TransmisisonControl constructor.

+++ src/transmissioncontrol.cpp
@@ -54,10 +54,28 @@ TransmissionControl::TransmissionControl() {
         qDebug() << "File " << tmp.fileName() << " has been parsed without error.";
     }
 
+
+
+    // initialise transmission object
     transmission = new QProcess(this);
-    transmission->start("/usr/bin/transmission-daemon",QStringList() << "-f"); // transmission must be run in foreground mode (-f flag) or it cannot be killed by qprocess
-    if (!transmission->waitForStarted())
-        qWarning() << "Transmission failed to start";
+
+    // when state of transmission process changes, update the transmissionOn variable using the updateTransmissionState slot
+    this->connect(
+                transmission, SIGNAL(stateChanged(QProcess::ProcessState)),
+                this, SLOT(updateTransmissionState(QProcess::ProcessState))
+                );
+
+    transmission->setProgram("transmission-daemon");
+    transmission->setArguments(QStringList() << "-f"); // transmission must be run in foreground mode (-f flag) or it cannot be killed by qprocess
+
+    // create application settings object and use it to decide whether to start the daemon
+    appSettings = new AppSettings;
+    if (appSettings->autoStartDaemon()) {
+        qDebug() << "Autostarting daemon due to user setting";
+        startTransmission();
+    } else {
+        qDebug() << "Not autostarting daemon due to user setting";
+    }
 
 
 }
@@ -65,31 +83,92 @@ TransmissionControl::TransmissionControl() {
 
 
 TransmissionControl::~TransmissionControl() {
+
     qDebug() << "Destroying TransmissionControl object";
-
-    // stop the transmission process when the app stops
-    transmission->terminate(); // send SIGTERM
-    // code below is probably unnecessary now, running transmission-daemon in foreground mode allows it to be terminated cleanly
-    if (!transmission->waitForFinished()) {
-        qWarning() << "couldn't kill process, sending SIGKILL";
-        transmission->kill(); // send SIGKILL
-        if (!transmission->waitForFinished()) {
-            qCritical() << "still couldn't kill process. Transmission may still be running!";
-        }
-    }
+    stopTransmission();
 
 }
 
 
-
-bool TransmissionControl::isTransmissionRunning() {
-    return 1;
+bool TransmissionControl::isTransmissionOn() {
+    return transmissionOn;
 }
 
-
+void TransmissionControl::updateTransmissionState(QProcess::ProcessState state) {
+    qDebug() << "update transmission state called";
+    if (state == QProcess::Running)
+        transmissionOn = true;
+    else
+        transmissionOn = false;
+    emit transmissionStateChanged();
+}
 
 void TransmissionControl::setTransmissionRunning(bool state) {
-    qDebug() << state;
+
+    if (state) {
+        qDebug() << "Setting transmission running";
+    } else {
+        qDebug() << "Stopping transmission";
+    }
+
+    if (state == 1) {
+        startTransmission();
+    } else {
+        stopTransmission();
+    }
 }
 
 
+void TransmissionControl::startTransmission() {
+
+    qDebug() << "startTransmission called";
+
+    // get current state
+    QProcess::ProcessState state = transmission->state();
+
+    // check if transmission is already running
+    if (state == QProcess::Running) {
+        qDebug() << "Transmission is already running, doing nothing.";
+    } else if (state == QProcess::Starting) {
+        qDebug() << "Transmission is already starting up, doing nothing.";
+    } else  if (state == QProcess::NotRunning) {
+        qDebug() << "Transmission is not running. Starting...";
+        transmission->start();
+        if (!transmission->waitForStarted())
+            qWarning() << "Transmission failed to start";
+    }
+}
+
+
+void TransmissionControl::stopTransmission() {
+
+    qDebug() << "stopTransmission called";
+    // get current state
+    QProcess::ProcessState state = transmission->state();
+
+    // check if transmission is already running
+    if (state == QProcess::NotRunning) {
+        qDebug() << "Transmission is not running, doing nothing.";
+    } else if (state == QProcess::Starting || state == QProcess::Running) {
+        qDebug() << "Transmission is running, stopping....";
+        transmission->terminate();
+        if (!transmission->waitForFinished()) {
+            qWarning() << "couldn't kill process, sending SIGKILL";
+            transmission->kill(); // send SIGKILL
+            if (!transmission->waitForFinished()) {
+                qCritical() << "still couldn't kill process. Transmission may still be running!";
+            }
+        }
+    }
+}
+
+AppSettings* TransmissionControl::getAppSettings() {
+    qDebug() << "Reading appSettings member object of transmission control from QML";
+    return appSettings;
+}
+
+void TransmissionControl::setAppSettings(AppSettings *settings) {
+    qDebug() << "Writing appSettings member object of transmission control from QML";
+    appSettings = settings;
+    emit appSettingsChanged();
+}
